Cleanup index.ts
